fix(cd-service): move locks to processBatch function
Ref: SRX-FIMM1N
